+++ src/main/java/com/example/tastefulai/domain/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.example.tastefulai.domain.member.service;
 
-import com.example.tastefulai.domain.image.dto.ProfileResponseDto;
+import com.example.tastefulai.domain.member.dto.ProfileResponseDto;
 import com.example.tastefulai.domain.member.dto.MemberResponseDto;
 import com.example.tastefulai.domain.member.entity.Member;
 import com.example.tastefulai.domain.member.enums.GenderRole;
